Move sample data into its own file and reseed
