+++ sampleData.js
@@ -1,0 +1,46 @@
+let sampleSeedData = [
+  {
+    name: "Sample Project One",
+    palettes: [
+      {
+        name: "Sample Palette one",
+        color_one: '#F7D951',
+        color_two: '#75B1FF',
+        color_three: '#985EEE',
+        color_four: '#5ED49B',
+        color_five: '#44638F',
+      },
+      {
+        name: "Sample Palette two",
+        color_one: '#DC5551',
+        color_two: '#F2AD00',
+        color_three: '#6CA131',
+        color_four: '#4B5DC6',
+        color_five: '#1F0E95',
+      }
+    ]
+  },
+  {
+    name: "Sample Project Two",
+    palettes: [
+      {
+        name: "Sample Palette one",
+        color_one: '#9BF14F',
+        color_two: '#145b9b',
+        color_three: '#F14F80',
+        color_four: '#ffad64',
+        color_five: '#ffdaf4',
+      },
+      {
+        name: "Sample Palette two",
+        color_one: '#00CFAA',
+        color_two: '#4E1C80',
+        color_three: '#0015e6',
+        color_four: '#3eff88',
+        color_five: '#B34E7E',
+      }
+    ]
+  }
+];
+
+module.exports = sampleSeedData;
